fix: pre-validation script fix and cleanup
